split worker file up to be distributed among model, train and worker
